Mariam_SWE/Mariam_SWE/requirement.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20
